pyproject.toml
[project]
name = "mfp"
version = "0.1.0"
description = "ModelFunctionProtocol — Turn any Swagger into LLM-native functions"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
keywords = ["mcp", "llm", "swagger", "openapi", "code-execution"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "docker>=7.0.0",
    "pyyaml>=6.0",
    "orjson>=3.10.0",
    "jinja2>=3.1.0",
    "ruff>=0.9.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.42.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
    "mypy>=1.14",
    "types-PyYAML>=6.0",
    "respx>=0.21.0",
]

[project.scripts]
mfp = "mfp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mfp"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["A002", "A003"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=mfp --cov-report=term-missing"

[tool.coverage.run]
source = ["src/mfp"]
omit = ["tests/*"]
